+++ apps/client/src/app/pages/portfolio/transactions/transactions-page.module.ts
@@ -4,6 +4,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RouterModule } from '@angular/router';
 import { GfTransactionsTableModule } from '@ghostfolio/client/components/transactions-table/transactions-table.module';
+import { ImportTransactionsService } from '@ghostfolio/client/services/import-transactions.service';
 
 import { GfCreateOrUpdateTransactionDialogModule } from './create-or-update-transaction-dialog/create-or-update-transaction-dialog.module';
 import { GfImportTransactionDialogModule } from './import-transaction-dialog/import-transaction-dialog.module';
@@ -23,7 +24,7 @@ import { TransactionsPageComponent } from './transactions-page.component';
     RouterModule,
     TransactionsPageRoutingModule
   ],
-  providers: [],
+  providers: [ImportTransactionsService],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class TransactionsPageModule {}