+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Extended the import functionality for transactions by `csv` files
+- Introduced the primary data source
+
+### Changed
+
+- Restricted the file selector of the import functionality for transactions to `csv` and `json`
+
 ## 1.60.0 - 13.10.2021
 
 ### Added